# n8n-playground/openapi_spec.json
{"openapi":"3.1.0","info":{"title":"n8n API Playground","description":"Comprehensive n8n API integration and workflow automation platform","version":"1.0.0"},"paths":{"/health":{"get":{"summary":"Health Check","description":"Health check endpoint.","operationId":"health_check_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/":{"get":{"summary":"Root","description":"Root endpoint with API information.","operationId":"root__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/workflow-automation/workflow-automation/workflows":{"post":{"tags":["Workflow Automation","Workflow Automation"],"summary":"Create a new workflow","description":"Create a new workflow from a template with custom parameters.","operationId":"create_workflow_workflow_automation_workflow_automation_workflows_post","requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateWorkflowRequest"}}}},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Workflow"}}}},"404":{"description":"Not found"},"422":{"description":"Validation error"},"500":{"description":"Internal server error"}}},"get":{"tags":["Workflow Automation","Workflow Automation"],"summary":"List workflows","description":"Get a paginated list of workflows with optional filtering.","operationId":"list_workflows_workflow_automation_workflow_automation_workflows_get","parameters":[{"name":"page","in":"query","required":false,"schema":{"type":"integer","minimum":1,"description":"Page number","default":1,"title":"Page"},"description":"Page number"},{"name":"page_size","in":"query","required":false,"schema":{"type":"integer","maximum":100,"minimum":1,"description":"Items per page","default":20,"title":"Page Size"},"description":"Items per page"},{"name":"status","in":"query","required":false,"schema":{"anyOf":[{"$ref":"#/components/schemas/WorkflowStatus"},{"type":"null"}],"description":"Filter by status","title":"Status"},"description":"Filter by status"},{"name":"template_name","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"description":"Filter by template name","title":"Template Name"},"description":"Filter by template name"},{"name":"tags","in":"query","required":false,"schema":{"anyOf":[{"type":"array","items":{"type":"string"}},{"type":"null"}],"description":"Filter by tags","title":"Tags"},"description":"Filter by tags"},{"name":"search","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"description":"Search in name and description","title":"Search"},"description":"Search in name and description"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/WorkflowListResponse"}}}},"404":{"description":"Not found"},"422":{"description":"Validation error"},"500":{"description":"Internal server error"}}}},"/workflow-automation/workflow-automation/workflows/{workflow_id}":{"get":{"tags":["Workflow Automation","Workflow Automation"],"summary":"Get workflow details","description":"Get detailed information about a specific workflow.","operationId":"get_workflow_workflow_automation_workflow_automation_workflows__workflow_id__get","parameters":[{"name":"workflow_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Workflow Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Workflow"}}}},"404":{"description":"Not found"},"422":{"description":"Validation error"},"500":{"description":"Internal server error"}}},"put":{"tags":["Workflow Automation","Workflow Automation"],"summary":"Update workflow","description":"Update workflow configuration and parameters.","operationId":"update_workflow_workflow_automation_workflow_automation_workflows__workflow_id__put","parameters":[{"name":"workflow_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Workflow Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateWorkflowRequest"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Workflow"}}}},"404":{"description":"Not found"},"422":{"description":"Validation error"},"500":{"description":"Internal server error"}}},"delete":{"tags":["Workflow Automation","Workflow Automation"],"summary":"Delete workflow","description":"Delete a workflow and all its executions.","operationId":"delete_workflow_workflow_automation_workflow_automation_workflows__workflow_id__delete","parameters":[{"name":"workflow_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Workflow Id"}},{"name":"force","in":"query","required":false,"schema":{"type":"boolean","description":"Force delete even if executions are running","default":false,"title":"Force"},"description":"Force delete even if executions are running"}],"responses":{"204":{"description":"Successful Response"},"404":{"description":"Not found"},"422":{"description":"Validation error"},"500":{"description":"Internal server error"}}}},"/workflow-automation/workflow-automation/workflows/{workflow_id}/execute":{"post":{"tags":["Workflow Automation","Workflow Automation"],"summary":"Execute workflow","description":"Start execution of a workflow with optional parameter overrides.","operationId":"execute_workflow_workflow_automation_workflow_automation_workflows__workflow_id__execute_post","parameters":[{"name":"workflow_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Workflow Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ExecuteWorkflowRequest"}}}},"responses":{"202":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/WorkflowExecution"}}}},"404":{"description":"Not found"},"422":{"description":"Validation error"},"500":{"description":"Internal server error"}}}},"/workflow-automation/workflow-automation/workflows/{workflow_id}/executions":{"get":{"tags":["Workflow Automation","Workflow Automation"],"summary":"List workflow executions","description":"Get a paginated list of executions for a specific workflow.","operationId":"list_workflow_executions_workflow_automation_workflow_automation_workflows__workflow_id__executions_get","parameters":[{"name":"workflow_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Workflow Id"}},{"name":"page","in":"query","required":false,"schema":{"type":"integer","minimum":1,"description":"Page number","default":1,"title":"Page"},"description":"Page number"},{"name":"page_size","in":"query","required":false,"schema":{"type":"integer","maximum":100,"minimum":1,"description":"Items per page","default":20,"title":"Page Size"},"description":"Items per page"},{"name":"status","in":"query","required":false,"schema":{"anyOf":[{"$ref":"#/components/schemas/ExecutionStatus"},{"type":"null"}],"description":"Filter by status","title":"Status"},"description":"Filter by status"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ExecutionListResponse"}}}},"404":{"description":"Not found"},"422":{"description":"Validation error"},"500":{"description":"Internal server error"}}}},"/workflow-automation/workflow-automation/executions/{execution_id}":{"get":{"tags":["Workflow Automation","Workflow Automation"],"summary":"Get execution details","description":"Get detailed information about a specific execution.","operationId":"get_execution_workflow_automation_workflow_automation_executions__execution_id__get","parameters":[{"name":"execution_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Execution Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/WorkflowExecution"}}}},"404":{"description":"Not found"},"422":{"description":"Validation error"},"500":{"description":"Internal server error"}}}},"/workflow-automation/workflow-automation/executions/{execution_id}/cancel":{"post":{"tags":["Workflow Automation","Workflow Automation"],"summary":"Cancel execution","description":"Cancel a running workflow execution.","operationId":"cancel_execution_workflow_automation_workflow_automation_executions__execution_id__cancel_post","parameters":[{"name":"execution_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Execution Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/WorkflowExecution"}}}},"404":{"description":"Not found"},"422":{"description":"Validation error"},"500":{"description":"Internal server error"}}}},"/workflow-automation/workflow-automation/workflows/{workflow_id}/stats":{"get":{"tags":["Workflow Automation","Workflow Automation"],"summary":"Get workflow statistics","description":"Get execution statistics and metrics for a workflow.","operationId":"get_workflow_stats_workflow_automation_workflow_automation_workflows__workflow_id__stats_get","parameters":[{"name":"workflow_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Workflow Id"}},{"name":"days","in":"query","required":false,"schema":{"type":"integer","maximum":365,"minimum":1,"description":"Number of days to include","default":30,"title":"Days"},"description":"Number of days to include"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/WorkflowStatsResponse"}}}},"404":{"description":"Not found"},"422":{"description":"Validation error"},"500":{"description":"Internal server error"}}}},"/workflow-automation/workflow-automation/health":{"get":{"tags":["Workflow Automation","Workflow Automation"],"summary":"Health check","description":"Check the health status of the workflow automation module.","operationId":"health_check_workflow_automation_workflow_automation_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"object","title":"Response Health Check Workflow Automation Workflow Automation Health Get"}}}},"404":{"description":"Not found"},"422":{"description":"Validation error"},"500":{"description":"Internal server error"}}}},"/workflow-automation/workflow-automation/templates":{"get":{"tags":["Workflow Automation","Workflow Automation"],"summary":"List workflow templates","description":"Get a list of available workflow templates.","operationId":"list_templates_workflow_automation_workflow_automation_templates_get","parameters":[{"name":"category","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"description":"Filter by category","title":"Category"},"description":"Filter by category"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"object","title":"Response List Templates Workflow Automation Workflow Automation Templates Get"}}}},"404":{"description":"Not found"},"422":{"description":"Validation error"},"500":{"description":"Internal server error"}}}},"/workflow-automation/workflow-automation/templates/{template_name}":{"get":{"tags":["Workflow Automation","Workflow Automation"],"summary":"Get template details","description":"Get detailed information about a specific template.","operationId":"get_template_workflow_automation_workflow_automation_templates__template_name__get","parameters":[{"name":"template_name","in":"path","required":true,"schema":{"type":"string","title":"Template Name"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"object","title":"Response Get Template Workflow Automation Workflow Automation Templates  Template Name  Get"}}}},"404":{"description":"Not found"},"422":{"description":"Validation error"},"500":{"description":"Internal server error"}}}},"/metrics":{"get":{"summary":"Metrics","operationId":"metrics_metrics_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}}},"components":{"schemas":{"CreateWorkflowRequest":{"properties":{"template_name":{"type":"string","title":"Template Name"},"name":{"type":"string","title":"Name"},"description":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Description"},"parameters":{"type":"object","title":"Parameters"},"tags":{"items":{"type":"string"},"type":"array","title":"Tags"}},"type":"object","required":["template_name","name"],"title":"CreateWorkflowRequest","description":"Request model for creating workflow."},"ExecuteWorkflowRequest":{"properties":{"parameters":{"anyOf":[{"type":"object"},{"type":"null"}],"title":"Parameters"},"wait_for_completion":{"type":"boolean","title":"Wait For Completion","default":false},"timeout":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Timeout"}},"type":"object","title":"ExecuteWorkflowRequest","description":"Request model for executing workflow."},"ExecutionListResponse":{"properties":{"executions":{"items":{"$ref":"#/components/schemas/WorkflowExecution"},"type":"array","title":"Executions"},"total":{"type":"integer","title":"Total"},"page":{"type":"integer","title":"Page"},"page_size":{"type":"integer","title":"Page Size"},"has_next":{"type":"boolean","title":"Has Next"},"has_previous":{"type":"boolean","title":"Has Previous"}},"type":"object","required":["executions","total","page","page_size","has_next","has_previous"],"title":"ExecutionListResponse","description":"Response model for execution list."},"ExecutionStatus":{"type":"string","enum":["pending","running","success","failed","cancelled","timeout"],"title":"ExecutionStatus","description":"Execution status enumeration."},"UpdateWorkflowRequest":{"properties":{"name":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Name"},"description":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Description"},"parameters":{"anyOf":[{"type":"object"},{"type":"null"}],"title":"Parameters"},"tags":{"anyOf":[{"items":{"type":"string"},"type":"array"},{"type":"null"}],"title":"Tags"},"status":{"anyOf":[{"$ref":"#/components/schemas/WorkflowStatus"},{"type":"null"}]}},"type":"object","title":"UpdateWorkflowRequest","description":"Request model for updating workflow."},"Workflow":{"properties":{"id":{"type":"string","title":"Id"},"name":{"type":"string","title":"Name"},"description":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Description"},"template_name":{"type":"string","title":"Template Name"},"template_version":{"type":"string","title":"Template Version"},"parameters":{"type":"object","title":"Parameters"},"workflow_data":{"type":"object","title":"Workflow Data"},"tags":{"items":{"type":"string"},"type":"array","title":"Tags"},"status":{"allOf":[{"$ref":"#/components/schemas/WorkflowStatus"}],"default":"created"},"created_at":{"type":"string","format":"date-time","title":"Created At"},"updated_at":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Updated At"},"last_executed":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Last Executed"},"execution_count":{"type":"integer","title":"Execution Count","default":0},"success_count":{"type":"integer","title":"Success Count","default":0},"failure_count":{"type":"integer","title":"Failure Count","default":0},"n8n_workflow_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"N8N Workflow Id"},"n8n_workflow_url":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"N8N Workflow Url"}},"type":"object","required":["name","template_name","template_version"],"title":"Workflow","description":"Workflow model."},"WorkflowExecution":{"properties":{"id":{"type":"string","title":"Id"},"workflow_id":{"type":"string","title":"Workflow Id"},"parameters":{"type":"object","title":"Parameters"},"status":{"allOf":[{"$ref":"#/components/schemas/ExecutionStatus"}],"default":"pending"},"created_at":{"type":"string","format":"date-time","title":"Created At"},"started_at":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Started At"},"completed_at":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Completed At"},"execution_time":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Execution Time"},"result":{"anyOf":[{"type":"object"},{"type":"null"}],"title":"Result"},"error":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Error"},"logs":{"items":{"type":"string"},"type":"array","title":"Logs"},"n8n_execution_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"N8N Execution Id"},"n8n_execution_url":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"N8N Execution Url"},"triggered_by":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Triggered By"},"trigger_data":{"anyOf":[{"type":"object"},{"type":"null"}],"title":"Trigger Data"}},"type":"object","required":["workflow_id"],"title":"WorkflowExecution","description":"Workflow execution model."},"WorkflowListResponse":{"properties":{"workflows":{"items":{"$ref":"#/components/schemas/Workflow"},"type":"array","title":"Workflows"},"total":{"type":"integer","title":"Total"},"page":{"type":"integer","title":"Page"},"page_size":{"type":"integer","title":"Page Size"},"has_next":{"type":"boolean","title":"Has Next"},"has_previous":{"type":"boolean","title":"Has Previous"}},"type":"object","required":["workflows","total","page","page_size","has_next","has_previous"],"title":"WorkflowListResponse","description":"Response model for workflow list."},"WorkflowStatsResponse":{"properties":{"total_workflows":{"type":"integer","title":"Total Workflows"},"active_workflows":{"type":"integer","title":"Active Workflows"},"total_executions":{"type":"integer","title":"Total Executions"},"successful_executions":{"type":"integer","title":"Successful Executions"},"failed_executions":{"type":"integer","title":"Failed Executions"},"average_execution_time":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Average Execution Time"},"workflows_by_status":{"additionalProperties":{"type":"integer"},"type":"object","title":"Workflows By Status"},"executions_by_status":{"additionalProperties":{"type":"integer"},"type":"object","title":"Executions By Status"},"recent_executions":{"items":{"$ref":"#/components/schemas/WorkflowExecution"},"type":"array","title":"Recent Executions"},"top_workflows":{"items":{"type":"object"},"type":"array","title":"Top Workflows"}},"type":"object","required":["total_workflows","active_workflows","total_executions","successful_executions","failed_executions","workflows_by_status","executions_by_status","recent_executions","top_workflows"],"title":"WorkflowStatsResponse","description":"Response model for workflow statistics."},"WorkflowStatus":{"type":"string","enum":["created","active","inactive","archived","error"],"title":"WorkflowStatus","description":"Workflow status enumeration."}}}}